requirements.txt
tensorflow-gpu
pillow
matplotlib
seaborn
requests
scipy